lab-02/requirements.txt
requests
pandas
python-dotenv
GitPython